Versioning the Graphql controller to eliminate the need to change the validator.
