--- lib/manageiq/api/common/application_controller_mixins/graphql.rb
@@ -1,17 +1,0 @@
-module ManageIQ
-  module API
-    module Common
-      module ApplicationControllerMixins
-        module Graphql
-          def api_version
-            "v#{::ManageIQ::API::Common::GraphQL.version(request)}"
-          end
-
-          def api_doc
-            @api_doc ||= ::ManageIQ::API::Common::OpenApi::Docs.instance[api_version[1..-1].sub(/x/, ".")]
-          end
-        end
-      end
-    end
-  end
-end
