+++ lib/manageiq/api/common/application_controller_mixins/request_body_validation.rb
@@ -42,24 +42,14 @@ module ManageIQ
             return unless request.post? || request.patch?
             return unless self.class.openapi_enabled
 
-            raw_api_version = try(:api_version) || self.class.send(:api_version)
-            api_version     = raw_api_version[1..-1].sub(/x/, ".")
+            api_version = self.class.send(:api_version)[1..-1].sub(/x/, ".")
 
-            if respond_to?(:api_doc)
-              api_doc.validate!(
-                request.method,
-                request.path,
-                api_version,
-                body_params.as_json
-              )
-            else
-              self.class.send(:api_doc).validate!(
-                request.method,
-                request.path,
-                api_version,
-                body_params.as_json
-              )
-            end
+            self.class.send(:api_doc).validate!(
+              request.method,
+              request.path,
+              api_version,
+              body_params.as_json
+            )
           rescue OpenAPIParser::OpenAPIError => exception
             error_document = ManageIQ::API::Common::ErrorDocument.new.add(400, exception.message)
             render :json => error_document.to_h, :status => :bad_request
