+++ spec/dummy/app/controllers/api/v1/graphql_controller.rb
@@ -1,0 +1,17 @@
+require "manageiq/api/common/graphql"
+
+module Api
+  module V1
+    class GraphqlController < ApplicationController
+      def query
+        graphql_api_schema = ::ManageIQ::API::Common::GraphQL::Generator.init_schema(request)
+        variables = ::ManageIQ::API::Common::GraphQL.ensure_hash(params[:variables])
+        result = graphql_api_schema.execute(
+          params[:query],
+          :variables => variables
+        )
+        render :json => result
+      end
+    end
+  end
+end
